+++ src/mesher/shared.ts
@@ -64,7 +64,7 @@ export interface MesherMainEvents {
   geometry: { type: 'geometry'; key: string; geometry: MesherGeometryOutput; workerIndex: number };
   sectionFinished: { type: 'sectionFinished'; key: string; workerIndex: number; processTime?: number };
   blockStateModelInfo: { type: 'blockStateModelInfo'; info: Record<string, BlockStateModelInfo> };
-  heightmap: { type: 'heightmap'; key: string; heightmap: Uint8Array };
+  heightmap: { type: 'heightmap'; key: string; heightmap: Int16Array };
 }
 
 export type MesherMainEvent = MesherMainEvents[keyof MesherMainEvents]
